+++ lib/context/context_extensions.dart
@@ -1,19 +1,13 @@
 part of '../flutter_extend.dart';
 
 extension ContextExtensions on BuildContext {
-  /// Get Screen Width
-  double get screenWidth => MediaQuery.sizeOf(this).width;
+  /// Hide Keyboard
+  void hideKeyboard() => FocusScope.of(this).unfocus();
 
-  /// Get Screen Height
-  double get screenHeight => MediaQuery.sizeOf(this).height;
-
-  /// Get Theme Data
-  ThemeData get theme => Theme.of(this);
-
-  /// Get Color Scheme
-  ColorScheme get colorScheme => Theme.of(this).colorScheme;
-
-  /// Check Dark Mode
-  bool get isDarkMode =>
-      MediaQuery.of(this).platformBrightness == Brightness.dark;
+  /// Perform Safe SetState operations by Checking whether the widget is disposed
+  void safeSetState(VoidCallback callback) {
+    if (mounted) {
+      callback();
+    }
+  }
 }